Extract deployment fixture to share between tests
