+++ subprojects/tooling-api/src/crossVersionTest/groovy/org/gradle/integtests/tooling/r25/DeploymentHandleContinuousBuildCrossVersionSpec.groovy
@@ -19,118 +19,40 @@ package org.gradle.integtests.tooling.r25
 import org.gradle.integtests.fixtures.executer.GradleVersions
 import org.gradle.integtests.tooling.fixture.ContinuousBuildToolingApiSpecification
 import org.gradle.integtests.tooling.fixture.TargetGradleVersion
+import org.gradle.test.fixtures.TestDeploymentFixture
 import spock.lang.Timeout
 
 @Timeout(120)
 @TargetGradleVersion(GradleVersions.SUPPORTS_DEPLOYMENT_REGISTRY)
 class DeploymentHandleContinuousBuildCrossVersionSpec extends ContinuousBuildToolingApiSpecification {
-    File triggerFile = file('triggerFile')
-    File keyFile = file('keyFile')
+    def fixture = new TestDeploymentFixture()
 
     def setup() {
-        buildFile << """
-            import javax.inject.Inject
-            import org.gradle.deployment.internal.Deployment
-            import org.gradle.deployment.internal.DeploymentHandle
-            import org.gradle.deployment.internal.DeploymentRegistry
-            import org.gradle.deployment.internal.DeploymentRegistry.ChangeBehavior
-
-            task runDeployment(type: RunTestDeployment) {
-                triggerFile = file('${triggerFile.name}')
-                keyFile = file('${keyFile.name}')
-            }
-
-            class TestDeploymentHandle implements DeploymentHandle {
-                final File keyFile
-                boolean running
-
-                @Inject 
-                TestDeploymentHandle(key, File keyFile) {
-                    this.keyFile = keyFile
-                    keyFile.text = key
-                }
-
-                public void start(Deployment deployment) {
-                    running = true
-                }
-                
-                public boolean isRunning() {
-                    return running
-                }
-
-                public void stop() {
-                    running = false
-                    keyFile.delete()
-                }
-            }
-
-            class RunTestDeployment extends DefaultTask {
-                @InputFile
-                File triggerFile
-
-                @OutputFile
-                File keyFile
-
-                @Inject
-                protected DeploymentRegistry getDeploymentRegistry() {
-                    throw new UnsupportedOperationException()
-                }
-
-                @TaskAction
-                void doDeployment() {
-                    // we set a different key for every build so we can detect if we
-                    // somehow get a different deployment handle between builds
-                    def key = System.currentTimeMillis()
-
-                    TestDeploymentHandle handle = getDeploymentRegistry().get('test', TestDeploymentHandle.class)
-                    if (handle == null) {
-                        // This should only happen once (1st build), so if we get a different value in keyFile between
-                        // builds then we know we can detect if we didn't get the same handle
-                        handle = getDeploymentRegistry().start('test', DeploymentRegistry.ChangeBehavior.NONE, TestDeploymentHandle.class, key, keyFile)
-                    }
-
-                    println "\\nCurrent Key: \$key, Deployed Key: \$handle.keyFile.text"
-                    assert handle.isRunning()
-                }
-            }
-        """
+        fixture.writeToProject(projectDir)
         buildTimeout = 30
     }
 
     def "deployment is stopped when continuous build is cancelled" () {
-        triggerFile.text = "0"
-
         when:
         runBuild(["runDeployment"]) {
             succeeds()
-            def key = file('keyFile').text
-            deploymentIsRunning(key)
+            def key = fixture.keyFile.text
+            fixture.assertDeploymentIsRunning(key)
 
-            waitBeforeModification triggerFile
-            triggerFile << "\n#a change"
+            waitBeforeModification fixture.triggerFile
+            fixture.triggerFile << "\n#a change"
             succeeds()
-            deploymentIsRunning(key)
+            fixture.assertDeploymentIsRunning(key)
 
-            waitBeforeModification triggerFile
-            triggerFile << "\n#another change"
+            waitBeforeModification fixture.triggerFile
+            fixture.triggerFile << "\n#another change"
             succeeds()
-            deploymentIsRunning(key)
+            fixture.assertDeploymentIsRunning(key)
 
             cancel()
         }
 
         then:
-        deploymentIsStopped()
-    }
-
-    void deploymentIsRunning(String key) {
-        // assert that the keyFile still exists and has the same contents (ie handle is still running)
-        assert keyFile.exists()
-        assert keyFile.text == key
-    }
-
-    void deploymentIsStopped() {
-        // assert that the deployment handle was stopped and removed the key file
-        assert !keyFile.exists()
+        fixture.assertDeploymentIsStopped()
     }
 }
